+++ src/imio/history/utils.py
@@ -8,24 +8,27 @@ from plone import api
 from Products.CMFPlone.utils import base_hasattr
 
 
-def getPreviousEvent(obj, event, checkMayView=True):
+def getPreviousEvent(obj, event, checkMayViewEvent=True, checkMayViewComment=True):
     '''Returns the previous event found in the history for the given p_event
        on p_obj if p_event is found.  p_checkMayView is passed to IImioHistory.getHistory
        and will enable/disable event's comments viewability check.'''
 
     adapter = getAdapter(obj, IImioHistory, 'workflow')
-    history = adapter.getHistory(checkMayView=checkMayView)
+    history = adapter.getHistory(
+        checkMayViewEvent=checkMayViewEvent, checkMayViewComment=checkMayViewComment)
     if event in history and history.index(event) > 0:
         return history[history.index(event) - 1]
 
 
-def getLastAction(obj, action='last', history_name='workflow', checkMayView=False):
+def getLastAction(obj, action='last', history_name='workflow',
+                  checkMayViewEvent=False, checkMayViewComment=False):
     '''Returns, from the p_history_name of p_obj, the last occurence of p_event.
        Default p_action is 'last' because we also want to be able to get
        an action that is 'None' in a particular p_history_name.'''
 
     adapter = getAdapter(obj, IImioHistory, history_name)
-    history = adapter.getHistory(checkMayView=checkMayView)
+    history = adapter.getHistory(
+        checkMayViewEvent=checkMayViewEvent, checkMayViewComment=checkMayViewComment)
 
     if action == 'last':
         # do not break if history is empty